requirements.txt
google-cloud-texttospeech==2.16.0
functions-framework==3.5.0
python-dotenv==1.0.1
